+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-Release v0.8.2 (2021-12-28)
+Release v0.8.3 (2021-12-28)
 ----------------------------
-First official release
+### Features & Improvements
+- Allow aliases to be abbreviated when invoked
+- Make cmd a non-mandatory CLI argument so that --init and --info can be used without a dummy command.
+
+Release v0.8.3 (2021-12-28)
+----------------------------
+First official release.